# api/openapi.yaml
openapi: 3.1.0
info:
  title: Sentinel REST API
  description: |
    Real-time indexing and event querying for Soroban contract events.

    ## API Stability & Compatibility Promise

    The `/v1/` REST API surface is **frozen and stable**. For the complete backward-compatibility guarantees, deprecation lifecycle, and versioning rules, see the [API Stability Policy](https://github.com/A4-Stellar/Sentinel/blob/dev/docs/API_STABILITY.md).

    ## Request validation

    Every endpoint validates its parameters before doing any work and answers a
    malformed request with `400` and the `ErrorResponse` envelope, where
    `error.code` is `INVALID_ARGUMENT` and `error.message` names the field that
    failed:

    ```json
    {"error": {"code": "INVALID_ARGUMENT", "message": "limit must be an integer between 1 and 200"}}
    ```

    Rules shared across endpoints:

    - `contractId` — Stellar contract strkey: `C` followed by 55 uppercase
      base32 characters.
    - path ids — canonical UUID v4.
    - ledger ranges — non-negative integers; the upper bound must be greater
      than or equal to the lower bound.
    - `limit` — an integer within the range documented for that endpoint.
    - `cursor` — an opaque cursor from a previous response; anything else is
      rejected rather than silently ignored.
    - `network` — `testnet` or `mainnet`.
    - Unknown query parameters are rejected, so a typo such as `?limitt=5`
...

    ## Pagination

...

...

    ```json
...
    ```

...

...
  version: 1.0.0
...
servers:
...
security:
  - ApiKeyAuth: []
paths:
...
  /v1/events:
    get:
      summary: List Soroban contract events
      description: Query events with optional filtering by contract, topics, and ledger range
      operationId: listEvents
      tags:
        - Events
      parameters:
        - name: contractId
          in: query
          schema:
            type: string
            pattern: "^C[A-Z2-7]{55}$"
          description: Filter by Soroban contract address (strkey format, 56 chars starting with C)
          example: CA7QYNF7SOWQ3GLR2BGMZEHXAVIRZA4KVWLTJJFC7MGXUA74P7UJVSGZ
        - name: topic0
          in: query
          schema:
            type: string
          description: Filter by primary topic (XDR-encoded)
        - name: topic1
          in: query
          schema:
            type: string
          description: Filter by secondary topic (XDR-encoded)
        - name: ledgerFrom
          in: query
          schema:
            type: integer
            minimum: 0
          description: Lower bound of ledger range (inclusive)
        - name: ledgerTo
          in: query
          schema:
            type: integer
            minimum: 0
          description: Upper bound of ledger range (inclusive)
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
          description: Maximum number of events to return
        - name: cursor
          in: query
          schema:
            type: string
          description: Opaque pagination cursor from previous response's next_cursor (for next page)
      responses:
        "200":
          description: List of events with pagination metadata
          headers:
            X-RateLimit-Limit:
              $ref: "#/components/headers/X-RateLimit-Limit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimitExceeded"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
  /v1/api-keys:
    post:
      summary: Create an API key
      description: Generate a new API key for authentication
      operationId: createApiKey
      tags:
        - API Keys
      security:
        - AdminKey: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                label:
                  type: string
                  description: Human-readable label for this key
                network:
                  type: string
                  enum: [testnet, mainnet]
                  default: mainnet
                  description: Network this key is associated with
                rate_limit_tier:
                  type: string
                  enum: [free, standard, premium, internal]
                  default: standard
                  description: Rate limit tier for this key
      responses:
        "201":
          description: API key created
          headers:
            Idempotent-Replayed:
              $ref: "#/components/headers/Idempotent-Replayed"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIKeyResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequestsIPOnly"

    get:
      summary: List API keys
      description: >-
        List API keys (admin only), keyset-paginated (issue #220). Ordered
        newest first (created_at DESC, id DESC as tiebreaker); cursor is
        opaque — never construct or parse it, just pass back the value
        returned in next_cursor.
      operationId: listApiKeys
      tags:
        - API Keys
      security:
        - AdminKey: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: cursor
          in: query
          required: false
          description: Opaque pagination cursor from a previous response's next_cursor.
          schema:
            type: string
      responses:
        "200":
          description: Page of API keys
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListAPIKeysResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequestsIPOnly"
  /v1/api-keys/{id}/rotate:
    post:
      summary: Rotate an API key
      description: >-
        Mint a replacement key that inherits the label, network, and
        rate-limit tier of the key identified by `id` (admin only). The
        plaintext key is returned once, in this response, and cannot be
        retrieved again. The existing key is not revoked by this call —
        revoke it with DELETE /v1/api-keys/{id} once callers have moved
        over, which evicts it from the auth cache immediately.
      operationId: rotateApiKey
      tags:
        - API Keys
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID of the active key to rotate
      responses:
        "201":
          description: Replacement key created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIKeyResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: No active API key with that id
        "429":
          $ref: "#/components/responses/TooManyRequestsIPOnly"
        "500":
          description: Rotation failed
  /v1/api-keys/{id}:
    patch:
      summary: Update an API key
      description: >-
        Update an API key's label and/or rate-limit tier (admin only). At
        least one of the two fields must be present. A tier change takes
        effect immediately — the shared tier cache is invalidated on success.
        Only active (non-revoked) keys can be updated.
      operationId: updateApiKey
      tags:
        - API Keys
      security:
        - AdminKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: API key ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                label:
                  type: string
                  description: New display label for the key
                rate_limit_tier:
                  type: string
                  description: New rate-limit tier name
      responses:
        "200":
          description: The updated API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIKeyResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Admin API key is not configured on the server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: API key not found (or already revoked)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: Request body exceeds the 1 MiB limit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          $ref: "#/components/responses/TooManyRequestsIPOnly"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
    delete:
      summary: Delete an API key
      description: Revoke and delete an API key (admin only)
      operationId: deleteApiKey
      tags:
        - API Keys
      security:
        - AdminKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: API key ID
      responses:
        "204":
          description: API key deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: API key not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          $ref: "#/components/responses/TooManyRequestsIPOnly"
...
components:
  schemas:
...
    EventListResponse:
      type: object
      additionalProperties: false
      required:
        - events
        - has_more
        - next_cursor
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/SorobanEvent"
          description: List of events
        has_more:
          type: boolean
          description: Whether more results are available
        next_cursor:
          type: string
          nullable: true
          description: Opaque cursor for next page (null if has_more is false)
...
    APIKeyResponse:
      type: object
      additionalProperties: false
      required:
        - id
        - key_prefix
        - label
        - network
        - rate_limit_tier
        - last_used_at
        - request_count
        - created_at
      properties:
        id:
          type: string
          format: uuid
        key_prefix:
          type: string
        key:
          type: string
          description: Raw key, returned only at creation time.
        label:
          type: string
        network:
          type: string
          enum: [testnet, mainnet]
        rate_limit_tier:
          type: string
        created_by:
          type: string
        last_used_at:
          type: string
          format: date-time
          nullable: true
        request_count:
          type: integer
          format: int64
        revoked_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
...
  headers:
    X-RateLimit-Limit:
      description: >-
        Requests allowed per window for this API key's rate-limit tier.
        Present on every response from an endpoint secured by ApiKeyAuth
        (2xx and 429 alike) once a valid X-API-Key was presented.
      required: true
      schema:
        type: integer
        minimum: 0
      example: 50
    X-RateLimit-Remaining:
      description: >-
        Requests remaining in the current window for this API key. 0 on the
        response that triggers a 429.
      required: true
      schema:
        type: integer
        minimum: 0
      example: 12
    X-RateLimit-Reset:
      description: >-
        Unix timestamp (seconds) when the current rate-limit window resets.
      required: true
      schema:
        type: integer
        format: int64
      example: 1732900000
    Retry-After:
      description: >-
        Seconds to wait before retrying. Present on 429 (rate limit
        exceeded, per-API-key or per-IP) and on 503 responses caused by the
        global concurrency cap shedding load. Not present on a 503 caused by
        an unavailable dependency (database/Redis/gRPC backend) — check the
        error envelope's `error.code` to distinguish the two.
      schema:
        type: integer
        minimum: 1
      example: 1
...
    Idempotent-Replayed:
      description: >-
        Present and set to "true" only when this response was replayed from
        a prior request carrying the same Idempotency-Key (issue #225).
        Absent on the original (first) execution.
      required: false
      schema:
        type: string
        enum: ["true"]
      example: "true"
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    RateLimitExceeded:
      description: >-
        Rate limit exceeded for this API key's tier (error.code
        RATE_LIMITED). Carries the same X-RateLimit-* headers as a
        successful response (X-RateLimit-Remaining is 0) plus Retry-After.
      headers:
        X-RateLimit-Limit:
          $ref: "#/components/headers/X-RateLimit-Limit"
        X-RateLimit-Remaining:
          $ref: "#/components/headers/X-RateLimit-Remaining"
        X-RateLimit-Reset:
          $ref: "#/components/headers/X-RateLimit-Reset"
        Retry-After:
          $ref: "#/components/headers/Retry-After"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    TooManyRequestsIPOnly:
      description: >-
        Per-IP rate limit exceeded (error.code RATE_LIMITED). Applies to
        endpoints not covered by per-API-key limiting (public endpoints, or
        admin endpoints authenticated via ADMIN_API_KEY rather than
        X-API-Key) — only Retry-After is set, no X-RateLimit-* headers.
      headers:
        Retry-After:
          $ref: "#/components/headers/Retry-After"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ServiceUnavailable:
      description: >-
        Service temporarily unavailable — either a dependency (database,
        Redis, gRPC backend) is down, or the server is shedding load under
        the global concurrency cap. When load-shedding is the cause,
        Retry-After is set; otherwise it is absent.
      headers:
        Retry-After:
          $ref: "#/components/headers/Retry-After"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: >-
        The supplied Idempotency-Key was already used with a request that
        does not match this one (error.code CONFLICT). Retry with a new key,
        or resend the exact original request body to get the original
        response replayed instead (issue #225).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: >-
        Opt-in de-duplication for retried create requests (issue #225). A
        retry carrying the same key and an identical request body replays
        the original response (see the `Idempotent-Replayed: true` response
        header) instead of creating a second resource. The same key reused
        with a different body is rejected with 409 Conflict. Keys are
        remembered for 24 hours; max length 255 characters.
      schema:
        type: string
        maxLength: 255
      example: 7da8f2b1-2e3c-4b8a-9f0d-6a1e5c3d9b7f
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication

    AdminKey:
      type: apiKey
      in: header
      name: X-Admin-Key
      description: Admin key for admin endpoints (environment variable)
...